requirements.txt
fastapi
uvicorn[standard]
httpx
questionary
requests
rich
jinja2
python-multipart
google-generativeai